# backend/docs/openapi/openapi.yaml
...
openapi: 3.1.0
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Login
      tags:
      - Auth Resource
  /auth/toggle/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: enabled
        in: query
        schema:
          type: boolean
      responses:
        "200":
          description: OK
      summary: Toggle User Status
      tags:
      - Auth Resource
  /bank-accounts:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Bank Account Controller
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccountRequestDTO"
        required: true
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Create
      tags:
      - Bank Account Controller
  /bank-accounts/account-info/agency:
    get:
      parameters:
      - name: account
        in: query
        schema:
          type: string
      - name: agency
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountRecipientInfoDTO"
      summary: Get By Agency And Account
      tags:
      - Bank Account Controller
  /bank-accounts/account-info/pix:
    get:
      parameters:
      - name: key
        in: query
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountRecipientInfoDTO"
      summary: Get By Pix Key
      tags:
      - Bank Account Controller
  /bank-accounts/by-account-number/{accountNumber}:
    get:
      parameters:
      - name: accountNumber
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      summary: Get By Account Number
      tags:
      - Bank Account Controller
  /bank-accounts/by-customer/{customerId}:
    get:
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get By Customer
      tags:
      - Bank Account Controller
  /bank-accounts/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get By Id
      tags:
      - Bank Account Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Delete
      tags:
      - Bank Account Controller
  /companies:
    get:
      summary: Listar empresas com paginação
      tags:
      - Companies
      parameters:
      - description: Número da página (começa em 0)
        name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - description: Quantidade de itens por página
        name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCompanyResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
    post:
      summary: Criar nova empresa conveniada
      tags:
      - Companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequestDTO"
        required: true
      responses:
        "201":
          description: Empresa criada com sucesso
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - company
  /companies/keycloak/{keycloakId}:
    get:
      summary: Buscar empresa por ID do Keycloak
      tags:
      - Companies
      parameters:
      - name: keycloakId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
  /companies/search:
    get:
      summary: "Buscar empresa por nome, CNPJ ou e-mail com paginação"
      tags:
      - Companies
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: query
        in: query
        schema:
          type: string
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCompanyResponseDTO"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
  /companies/{id}:
    put:
      summary: Atualizar dados da empresa
      tags:
      - Companies
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyRequestDTO"
        required: true
      responses:
        "200":
          description: Empresa atualizada
        "404":
          description: Empresa não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      security:
      - SecurityScheme:
        - admin
        - company
    get:
      summary: Buscar empresa por ID
      tags:
      - Companies
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Empresa encontrada
        "404":
          description: Empresa não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
    delete:
      summary: Remover empresa do sistema
      tags:
      - Companies
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Empresa removida
        "404":
          description: Empresa não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
...
  /customers/company/{keycloakId}:
    get:
      summary: Buscar clientes por ID da empresa
      tags:
      - Customers
      parameters:
      - name: keycloakId
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Clientes encontrados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResultCustomerResponseDTO"
        "404":
          description: Empresa não encontrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - admin
        - company
...
  /pix-keys/account/{accountId}:
    get:
      parameters:
      - name: accountId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
      summary: Get All By Account
      tags:
      - Pix Key Controller
...
components:
  schemas:
    AccountRecipientInfoDTO:
      type: object
      properties:
        name:
          type: string
        agencyNumber:
          type: string
        accountNumber:
          type: string
        pixKey:
          type: string
        pixKeyType:
          type: string
    BankAccountRequestDTO:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        accountNumber:
          type: string
        agencyNumber:
          type: string
    CompanyRequestDTO:
      type: object
      required:
      - name
      - cnpj
      - responsible
      - email
      - phone
      - type
      properties:
        name:
          type: string
          pattern: \S
        cnpj:
          type: string
          pattern: "\\d{14}"
        responsible:
          type: string
          pattern: \S
        email:
          type: string
          pattern: \S
        phone:
          type: string
          pattern: \S
        type:
          type: string
          pattern: \S
    CompanyResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        keycloakId:
          type: string
        name:
          type: string
        cnpj:
          type: string
        responsible:
          type: string
        email:
          type: string
        type:
          type: string
        phone:
          type: string
...
    CustomerResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        keycloakId:
          type: string
        enabled:
          type: boolean
        name:
          type: string
        cpf:
          type: string
        email:
          type: string
        phone:
          type: string
        birthDate:
          $ref: "#/components/schemas/LocalDate"
        companyId:
          type: integer
          format: int64
        companyName:
          type: string
        occupation:
          type: string
        street:
          type: string
        streetNumber:
          type: number
        city:
          type: string
        neighborhood:
          type: string
        state:
          type: string
        zipCode:
          type: string
        country:
          type: string
        createdBy:
          type: string
        updatedBy:
          type: string
        createdAt:
          $ref: "#/components/schemas/LocalDate"
        updatedAt:
          $ref: "#/components/schemas/LocalDate"
        virtualAccountNumber:
          type: string
        virtualAgencyNumber:
          type: string
        virtualBalance:
          type: number
        salary:
          type: number
...
    LocalDate:
      type: string
      format: date
      examples:
      - 2022-03-10
...
    PagedResultCompanyResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CompanyResponseDTO"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    PagedResultCustomerResponseDTO:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomerResponseDTO"
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://localhost:8080/realms/crediflow/.well-known/openid-configuration
      description: Authentication
...
info:
  title: backend API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8082
  description: Auto generated value
- url: http://0.0.0.0:8082
  description: Auto generated value