# backend/src/main/java/com/crediflow/dto/CompanyResponseDTO.java
package com.crediflow.dto;

public class CompanyResponseDTO {
    public Long id;
    public String keycloakId;
    public String name;
    public String cnpj;
    public String responsible;
    public String email;
    public String type;
    public String phone;
}